requirements.txt
flask==3.0.0
psycopg[binary]==3.2.12
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
